+++ Source/Core/EllipsoidalOccluder.js
@@ -150,5 +150,164 @@ define([
         return !isOccluded;
     };
 
+    /**
+     * Computes a point that can be used for horizon culling from a list of positions.  If the point is below
+     * the horizon, all of the positions are guaranteed to be below the horizon as well.  The returned point
+     * is expressed in the ellipsoid-scaled space and is suitable for use with
+     * {@link EllipsoidalOccluder#isScaledSpacePointVisible}.
+     *
+     * @param {Cartesian3} directionToPoint The direction that the computed point will lie along.
+     *                     A reasonable direction to use is the direction from the center of the ellipsoid to
+     *                     the center of the bounding sphere computed from the positions.  The direction need not
+     *                     be normalized.
+     * @param {Cartesian3[]} positions The positions from which to compute the horizon culling point.  The positions
+     *                       must be expressed in a reference frame centered at the ellipsoid and aligned with the
+     *                       ellipsoid's axes.
+     * @param {Cartesian3} [result] The instance on which to store the result instead of allocating a new instance.
+     * @returns {Cartesian3} The computed horizon culling point, expressed in the ellipsoid-scaled space.
+     */
+    EllipsoidalOccluder.prototype.computeHorizonCullingPoint = function(directionToPoint, positions, result) {
+        if (typeof directionToPoint === 'undefined') {
+            throw new DeveloperError('directionToPoint is required');
+        }
+        if (typeof positions === 'undefined') {
+            throw new DeveloperError('positions is required');
+        }
+
+        var ellipsoid = this._ellipsoid;
+
+        var scaledSpaceDirectionToPoint = computeScaledSpaceDirectionToPoint(ellipsoid, directionToPoint);
+
+        var resultMagnitude = 0.0;
+
+        for (var i = 0, len = positions.length; i < len; ++i) {
+            var position = positions[i];
+            var candidateMagnitude = computeMagnitude(ellipsoid, position, scaledSpaceDirectionToPoint);
+            resultMagnitude = Math.max(resultMagnitude, candidateMagnitude);
+        }
+
+        return magnitudeToPoint(scaledSpaceDirectionToPoint, resultMagnitude, result);
+    };
+
+    var positionScratch = new Cartesian3();
+
+    /**
+     * Computes a point that can be used for horizon culling from a list of positions.  If the point is below
+     * the horizon, all of the positions are guaranteed to be below the horizon as well.  The returned point
+     * is expressed in the ellipsoid-scaled space and is suitable for use with
+     * {@link EllipsoidalOccluder#isScaledSpacePointVisible}.
+     *
+     * @param {Cartesian3} directionToPoint The direction that the computed point will lie along.
+     *                     A reasonable direction to use is the direction from the center of the ellipsoid to
+     *                     the center of the bounding sphere computed from the positions.  The direction need not
+     *                     be normalized.
+     * @param {Number[]} vertices  The vertices from which to compute the horizon culling point.  The positions
+     *                   must be expressed in a reference frame centered at the ellipsoid and aligned with the
+     *                   ellipsoid's axes.
+     * @param {Number} [stride=3]
+     * @param {Cartesian3} [center=Cartesian3.ZERO]
+     * @param {Cartesian3} [result] The instance on which to store the result instead of allocating a new instance.
+     * @returns {Cartesian3} The computed horizon culling point, expressed in the ellipsoid-scaled space.
+     */
+    EllipsoidalOccluder.prototype.computeHorizonCullingPointFromVertices = function(directionToPoint, vertices, stride, center, result) {
+        if (typeof directionToPoint === 'undefined') {
+            throw new DeveloperError('directionToPoint is required');
+        }
+        if (typeof vertices === 'undefined') {
+            throw new DeveloperError('vertices is required');
+        }
+        if (typeof stride === 'undefined') {
+            throw new DeveloperError('stride is required');
+        }
+
+        center = defaultValue(center, Cartesian3.ZERO);
+
+        var ellipsoid = this._ellipsoid;
+
+        var scaledSpaceDirectionToPoint = computeScaledSpaceDirectionToPoint(ellipsoid, directionToPoint);
+
+        var resultMagnitude = 0.0;
+
+        for (var i = 0, len = vertices.length; i < len; i += stride) {
+            positionScratch.x = vertices[i] + center.x;
+            positionScratch.y = vertices[i + 1] + center.y;
+            positionScratch.z = vertices[i + 2] + center.z;
+
+            var candidateMagnitude = computeMagnitude(ellipsoid, positionScratch, scaledSpaceDirectionToPoint);
+            resultMagnitude = Math.max(resultMagnitude, candidateMagnitude);
+        }
+
+        return magnitudeToPoint(scaledSpaceDirectionToPoint, resultMagnitude, result);
+    };
+
+    var subsampleScratch = [];
+
+    /**
+     * Computes a point that can be used for horizon culling of an extent.  If the point is below
+     * the horizon, the ellipsoid-conforming extent is guaranteed to be below the horizon as well.
+     * The returned point is expressed in the ellipsoid-scaled space and is suitable for use with
+     * {@link EllipsoidalOccluder#isScaledSpacePointVisible}.
+     *
+     * @param {Extent} extent The extent for which to compute the horizon culling point.
+     * @param {Ellipsoid} ellipsoid The ellipsoid on which the extent is defined.  This may be different from
+     *                    the ellipsoid used by this instance for occlusion testing.
+     * @param {Cartesian3} [result] The instance on which to store the result instead of allocating a new instance.
+     * @returns {Cartesian3} The computed horizon culling point, expressed in the ellipsoid-scaled space.
+     */
+    EllipsoidalOccluder.prototype.computeHorizonCullingPointFromExtent = function(extent, ellipsoid, result) {
+        if (typeof extent === 'undefined') {
+            throw new DeveloperError('extent is required.');
+        }
+
+        var positions = extent.subsample(ellipsoid, subsampleScratch);
+        var bs = BoundingSphere.fromPoints(positions);
+
+        // If the bounding sphere center is too close to the center of the occluder, it doesn't make
+        // sense to try to horizon cull it.
+        if (bs.center.magnitude() < 0.1 * ellipsoid.getMinimumRadius()) {
+            return undefined;
+        }
+
+        return this.computeHorizonCullingPoint(bs.center, positions, result);
+    };
+
+    var scaledSpaceScratch = new Cartesian3();
+    var directionScratch = new Cartesian3();
+
+    function computeMagnitude(ellipsoid, position, scaledSpaceDirectionToPoint) {
+        var scaledSpacePosition = ellipsoid.transformPositionToScaledSpace(position, scaledSpaceScratch);
+        var magnitudeSquared = scaledSpacePosition.magnitudeSquared();
+        var magnitude = Math.sqrt(magnitudeSquared);
+        var direction = scaledSpacePosition.divideByScalar(magnitude, directionScratch);
+
+        // For the purpose of this computation, points below the ellipsoid are consider to be on it instead.
+        magnitudeSquared = Math.max(1.0, magnitudeSquared);
+        magnitude = Math.max(1.0, magnitude);
+
+        var cosAlpha = direction.dot(scaledSpaceDirectionToPoint);
+        var sinAlpha = direction.cross(scaledSpaceDirectionToPoint).magnitude();
+        var cosBeta = 1.0 / magnitude;
+        var sinBeta = Math.sqrt(magnitudeSquared - 1.0) * cosBeta;
+
+        return 1.0 / (cosAlpha * cosBeta - sinAlpha * sinBeta);
+    }
+
+    function magnitudeToPoint(scaledSpaceDirectionToPoint, resultMagnitude, result) {
+        // The horizon culling point is undefined if there were no positions from which to compute it,
+        // the directionToPoint is pointing opposite all of the positions,  or if we computed NaN or infinity.
+        if (resultMagnitude <= 0.0 || resultMagnitude === 1.0 / 0.0 || resultMagnitude !== resultMagnitude) {
+            return undefined;
+        }
+
+        return scaledSpaceDirectionToPoint.multiplyByScalar(resultMagnitude, result);
+    }
+
+    var directionToPointScratch = new Cartesian3();
+
+    function computeScaledSpaceDirectionToPoint(ellipsoid, directionToPoint) {
+        ellipsoid.transformPositionToScaledSpace(directionToPoint, directionToPointScratch);
+        return directionToPointScratch.normalize(directionToPointScratch);
+    }
+
     return EllipsoidalOccluder;
 });